# api/openapi.yaml
openapi: 3.0.3
info:
  title: Arc Issue Tracker API
  description: |
    Central issue tracking server for AI-assisted coding workflows.
    Arc provides project-scoped issue management with dependency tracking,
    labels, comments, and audit trails.
  version: 1.0.0
  contact:
    name: SentioLabs
    url: https://github.com/sentiolabs/arc
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"

    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      responses:
        "204":
          description: Project deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      description: Project ID
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
...
    Project:
      type: object
      required:
        - id
        - name
        - prefix
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique project ID (e.g., "proj-a1b2")
        name:
          type: string
          description: Display name
          maxLength: 100
        path:
          type: string
          description: Optional associated directory path
        description:
          type: string
          description: Project description
        prefix:
          type: string
          description: Issue ID prefix (e.g., "bd")
          maxLength: 10
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          description: New display name
          maxLength: 100
        path:
          type: string
          description: New associated directory path
        description:
          type: string
          description: New description
...